refactor: route remaining external storage boundaries
