+++ crates/e2e_test/src/lib.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 mod reliant;
+mod storage_api;
 
 // Common utilities for all E2E tests
 #[cfg(test)]
