+++ crates/e2e_test/src/reliant/node_interact_test.rs
@@ -14,10 +14,11 @@
 // limitations under the License.
 
 use crate::common::workspace_root;
+use crate::storage_api::{
+    TonicInterceptor, VolumeInfo, WalkDirOptions, gen_tonic_signature_interceptor, node_service_time_out_client,
+};
 use futures::future::join_all;
 use rmp_serde::{Deserializer, Serializer};
-use rustfs_ecstore::api::disk::{VolumeInfo, WalkDirOptions};
-use rustfs_ecstore::api::rpc::{TonicInterceptor, gen_tonic_signature_interceptor, node_service_time_out_client};
 use rustfs_filemeta::{MetaCacheEntry, MetacacheReader, MetacacheWriter};
 use rustfs_protos::proto_gen::node_service::WalkDirRequest;
 use rustfs_protos::{
