+++ crates/iam/src/lib.rs
@@ -15,19 +15,6 @@
 use crate::error::{Error, Result};
 use manager::IamCache;
 use oidc::OidcSys;
-use rustfs_ecstore::api::config::RUSTFS_CONFIG_PREFIX as ECSTORE_RUSTFS_CONFIG_PREFIX;
-use rustfs_ecstore::api::config::com::{
-    delete_config as ecstore_delete_config, read_config_no_lock as ecstore_read_config_no_lock,
-    read_config_with_metadata as ecstore_read_config_with_metadata, save_config as ecstore_save_config,
-    save_config_with_opts as ecstore_save_config_with_opts,
-};
-use rustfs_ecstore::api::error::{
-    Error as EcstoreErrorType, Result as EcstoreResultType, StorageError as EcstoreStorageError,
-    classify_system_path_failure_reason as ecstore_classify_system_path_failure_reason,
-};
-use rustfs_ecstore::api::global::is_first_cluster_node_local as ecstore_is_first_cluster_node_local;
-use rustfs_ecstore::api::notification::NotificationPeerErr as EcstoreNotificationPeerErr;
-use rustfs_ecstore::api::storage::ECStore as EcstoreStore;
 use std::sync::{Arc, OnceLock};
 use store::object::ObjectStore;
 use sys::IamSys;
@@ -48,66 +35,26 @@ pub mod oidc_state;
 mod root_credentials;
 mod runtime_sources;
 mod server_config;
+mod storage_api;
 pub mod store;
 pub mod sys;
 pub mod utils;
-
-pub(crate) const IAM_CONFIG_ROOT_PREFIX: &str = ECSTORE_RUSTFS_CONFIG_PREFIX;
-
-pub(crate) type IamEcstoreError = EcstoreErrorType;
-pub(crate) type IamStorageError = EcstoreStorageError;
-pub(crate) type IamStorageResult<T> = EcstoreResultType<T>;
-pub(crate) type IamStore = EcstoreStore;
-pub(crate) type IamConfigObjectInfo = <IamStore as rustfs_storage_api::ObjectOperations>::ObjectInfo;
-pub(crate) type IamConfigObjectOptions = <IamStore as rustfs_storage_api::ObjectOperations>::ObjectOptions;
+pub(crate) use storage_api::{
+    IAM_CONFIG_ROOT_PREFIX, IamEcstoreError, IamStorageError, IamStore, classify_iam_system_path_failure_reason,
+    delete_iam_config, is_iam_first_cluster_node_local, read_iam_config_no_lock, read_iam_config_with_metadata, save_iam_config,
+    save_iam_config_with_opts,
+};
 
 pub fn is_root_access_key(access_key: &str) -> bool {
     root_credentials::is_root_access_key(access_key)
-}
-
-pub(crate) async fn read_iam_config_no_lock(api: Arc<IamStore>, file: &str) -> IamStorageResult<Vec<u8>> {
-    ecstore_read_config_no_lock(api, file).await
-}
-
-pub(crate) async fn read_iam_config_with_metadata(
-    api: Arc<IamStore>,
-    file: &str,
-    opts: &IamConfigObjectOptions,
-) -> IamStorageResult<(Vec<u8>, IamConfigObjectInfo)> {
-    ecstore_read_config_with_metadata(api, file, opts).await
-}
-
-pub(crate) async fn save_iam_config(api: Arc<IamStore>, file: &str, data: Vec<u8>) -> IamStorageResult<()> {
-    ecstore_save_config(api, file, data).await
-}
-
-pub(crate) async fn save_iam_config_with_opts(
-    api: Arc<IamStore>,
-    file: &str,
-    data: Vec<u8>,
-    opts: &IamConfigObjectOptions,
-) -> IamStorageResult<()> {
-    ecstore_save_config_with_opts(api, file, data, opts).await
-}
-
-pub(crate) async fn delete_iam_config(api: Arc<IamStore>, file: &str) -> IamStorageResult<()> {
-    ecstore_delete_config(api, file).await
-}
-
-pub(crate) fn classify_iam_system_path_failure_reason(err: &IamEcstoreError) -> &'static str {
-    ecstore_classify_system_path_failure_reason(err)
-}
-
-pub(crate) async fn is_iam_first_cluster_node_local() -> bool {
-    ecstore_is_first_cluster_node_local().await
 }
 
 pub(crate) struct IamNotificationPeerErr {
     pub(crate) err: Option<IamEcstoreError>,
 }
 
-impl From<EcstoreNotificationPeerErr> for IamNotificationPeerErr {
-    fn from(value: EcstoreNotificationPeerErr) -> Self {
+impl From<storage_api::IamEcstoreNotificationPeerErr> for IamNotificationPeerErr {
+    fn from(value: storage_api::IamEcstoreNotificationPeerErr) -> Self {
         Self { err: value.err }
     }
 }
