+++ crates/notify/src/lib.rs
@@ -18,15 +18,6 @@
 //! It supports sending events to various targets
 //! (like Webhook and MQTT) and includes features like event persistence and retry on failure.
 
-use std::sync::Arc;
-
-use rustfs_ecstore::api::config::com::{
-    read_config_without_migrate as read_notify_config_without_migrate_from_backend,
-    save_server_config as save_notify_server_config_to_backend,
-};
-use rustfs_ecstore::api::global::resolve_object_store_handle as resolve_notify_object_store_handle_from_backend;
-pub(crate) use rustfs_ecstore::api::storage::ECStore as NotifyStore;
-
 mod bucket_config_manager;
 mod config_manager;
 mod error;
@@ -44,6 +35,7 @@ mod runtime_facade;
 mod runtime_view;
 mod services;
 mod status_view;
+mod storage_api;
 
 pub use bucket_config_manager::NotifyBucketConfigManager;
 pub use config_manager::{NotifyConfigManager, runtime_target_id_for_subsystem};
@@ -61,24 +53,6 @@ pub use runtime_facade::NotifyRuntimeFacade;
 pub use runtime_view::NotifyRuntimeView;
 pub use services::NotifyServices;
 pub use status_view::NotifyStatusView;
-
-pub(crate) fn resolve_notify_object_store_handle() -> Option<Arc<NotifyStore>> {
-    resolve_notify_object_store_handle_from_backend()
-}
-
-pub(crate) async fn read_notify_server_config_without_migrate(
-    store: Arc<NotifyStore>,
-) -> Result<rustfs_config::server_config::Config, String> {
-    read_notify_config_without_migrate_from_backend(store)
-        .await
-        .map_err(|err| err.to_string())
-}
-
-pub(crate) async fn save_notify_server_config(
-    store: Arc<NotifyStore>,
-    config: &rustfs_config::server_config::Config,
-) -> Result<(), String> {
-    save_notify_server_config_to_backend(store, config)
-        .await
-        .map_err(|err| err.to_string())
-}
+pub(crate) use storage_api::{
+    read_notify_server_config_without_migrate, resolve_notify_object_store_handle, save_notify_server_config,
+};
